Extract engine audio normalization operation
normalize_audio is a focused loudness-normalization operation around the FFmpeg loudnorm filter. This moves it into engine_audio_normalize.py while re-exporting normalize_audio from mcp_video.engine for compatibility.

Constraint: Preserve target LUFS validation, loudnorm filter behavior, audio bitrate, video copy behavior, and EditResult output.

Rejected: Move audio_waveform with normalize_audio | waveform analysis has a different result contract and failure behavior.

Confidence: high

Scope-risk: narrow

Directive: Keep audio analysis/filter operations split by behavior and explicitly escape FFmpeg filter values.

Tested: ruff check mcp_video/engine.py mcp_video/engine_audio_normalize.py

Tested: ruff format --check mcp_video/engine.py mcp_video/engine_audio_normalize.py

Tested: /opt/homebrew/bin/python3 normalize_audio re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine_advanced.py::TestNormalizeAudio tests/test_cli.py::TestCLINormalizeAudio tests/test_server.py::TestVideoNormalizeAudioTool tests/test_client.py::TestClientNormalizeAudio -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
